+++ packages/redux_persist_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `0.8.1` - 2019-03-03
+
+* Add example.
+
 ## `0.8.0` - 2018-11-27
 
 * Release v0.8.0. Last v0.x release line.